+++ minimax/TwoPlayerGame.java
@@ -2,8 +2,8 @@ package minimax;
 
 public interface TwoPlayerGame {
 
-    public int point(Table t, short player);
-    public Table turn(Table t, short player, Move move);
-    public Move[] possibleMoves(Table t, short player);
-    public boolean hasPossibleMove(Table t, short player);
+    public int point(Table t, byte player);
+    public Table turn(Table t, byte player, Move move);
+    public Move[] possibleMoves(Table t, byte player);
+    public boolean hasPossibleMove(Table t, byte player);
 }
